09-Celery,travis ci

+++ common/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class LogRecord(models.Model):
     method = models.CharField(max_length=200)
     path = models.CharField(max_length=200)
-    execution_time = models.FloatField()
+    execution_time = models.DecimalField(max_digits=10, decimal_places=6)
     created = models.DateTimeField(auto_now_add=True)
